Update Ktor version and modify dependency declarations
Updated the Ktor version from 2.3.4 to 2.3.5 in NetworkVersions.kt and versions.json. Also, shifted the dependencies from 'implementation' to 'api' to ensure that they are exposed to the compile classpath of modules that consume these libraries. This ensures better interoperability. Removed the 'kotlin.native.cacheKind=none' from project_gradle.properties.ft as it is no longer necessary.